BACKEND/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1
django-cors-headers==4.4.0
django-daraja==1.3.0
django-mpesa==2.0.14
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
packaging==24.1
phonenumbers==8.13.43
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
stripe==10.10.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
